Person view is now tabbed
