+++ common/messages/en/app.php
@@ -232,6 +232,8 @@ return [
     'PERFORMED_ACTION_UPDATE' => 'updated',
     'PERSON_BUTTON_CREATE' => 'Add person',
     'PERSON_DATA' => 'Person data',
+    'PERSON_DESCRIPTIONS_TAB' => 'Descriptions',
+    'PERSON_GM_TAB' => 'GM data',
     'PERSON_ID' => 'Person ID',
     'PERSON_KEY' => 'Key',
     'PERSON_NAME' => 'Person name',
@@ -241,7 +243,6 @@ return [
     'PERSON_VISIBILITY_GM' => 'To GM only',
     'PERSON_VISIBILITY_LOGGED' => 'To known users',
     'PERSON_VISIBILITY_NONE' => 'To none',
-    'PERSON_VISIBLE' => '',
     'PLACEHOLDER_NOT_YET_IMPLEMENTED' => 'Not yet implemented. This feature will appear in one of the future releases.',
     'PROMPT_SELECT_TYPE' => 'Select type ...',
     'PROMPT_USER' => 'Select user ...',
