+++ common/messages/pl/app.php
@@ -232,6 +232,8 @@ return [
     'PERFORMED_ACTION_UPDATE' => 'zmodyfikował[a]',
     'PERSON_BUTTON_CREATE' => 'Dodaj osobę',
     'PERSON_DATA' => 'Dane postaci',
+    'PERSON_DESCRIPTIONS_TAB' => 'Opisy',
+    'PERSON_GM_TAB' => 'Dane MG',
     'PERSON_ID' => 'ID postaci',
     'PERSON_KEY' => 'Klucz',
     'PERSON_NAME' => 'Imię osoby',
@@ -241,7 +243,6 @@ return [
     'PERSON_VISIBILITY_GM' => 'Dla MG',
     'PERSON_VISIBILITY_LOGGED' => 'Dla zalogowanych',
     'PERSON_VISIBILITY_NONE' => 'Dla nikogo',
-    'PERSON_VISIBLE' => '',
     'PLACEHOLDER_NOT_YET_IMPLEMENTED' => 'Nie zaimplementowano. Ta funkcjonalność pojawi się w jednej z przyszłych wersji.',
     'PROMPT_SELECT_TYPE' => 'Wybierz typ ...',
     'PROMPT_USER' => 'Wybierz użytkownika ...',
