+++ frontend/views/person/_view_descriptions_widget.php
@@ -1,0 +1,37 @@
+<?php
+/* @var $this yii\web\View */
+/* @var $model common\models\Person */
+/* @var $showPrivates bool */
+?>
+
+<?php if ($model->descriptionPack): ?>
+    <div id="descriptions">
+
+        <?= \yii\widgets\ListView::widget([
+            'dataProvider' => new \yii\data\ActiveDataProvider([
+                'query' => $model->descriptionPack->getDescriptionsInLanguageOfTheActiveUser(),
+                'sort' => ['defaultOrder' => ['position' => SORT_ASC]]
+            ]),
+            'itemOptions' => ['class' => 'item'],
+            'summary' => '',
+            'itemView' => function (\common\models\Description $model, $key, $index, $widget) {
+                return $this->render(
+                    '_view_descriptions',
+                    [
+                        'model' => $model,
+                        'key' => $key,
+                        'index' => $index,
+                        'widget' => $widget,
+                        'showPrivates' => $this->params['showPrivates']
+                    ]
+                );
+            },
+        ]) ?>
+
+        <div class="clearfix"></div>
+
+    </div>
+<?php else: ?>
+    <p><?= Yii::t('app', 'DESCRIPTIONS_NOT_FOUND'); ?></p>
+<?php endif; ?>
+
